==> pipeline/preparation/de/02--MdB-Posts-Processing.ipynb <==
import typing

import pandas
import seaborn

import cltrier_lib

import twon_agents

SEED: int = 1
DATA_PATH: str = "../../../data"

METRICS: typing.List[str] = [
    "public_metrics.impression_count",
    "public_metrics.like_count",
    "public_metrics.retweet_count",
    "public_metrics.reply_count",
]

posts: pandas.DataFrame = twon_agents.data.filter_tweets(
    pandas.read_csv(
        f"{DATA_PATH}/raw/twitter.german/GermanyMdBTweets_2023.csv", index_col=0
    )
    .merge(
        right=pandas.read_json(f"{DATA_PATH}/raw/twitter.german/mdb_meta.json"),
        left_on="username",
        right_on="twitter_username",
    )
)
posts.to_csv(f"{DATA_PATH}/interim/twitter.german.posts.csv")
posts

# calculate the distribution of replies
post_reply_dist: pandas.Series = (
    posts["public_metrics.reply_count"]
    .value_counts()
    .describe(percentiles=[0.25, 0.5, 0.75, 0.8, 0.85, 0.9, 0.95])
)
post_reply_dist

# generate a subset containing only the 90% most commented tweets
posts = posts[posts["public_metrics.reply_count"] > post_reply_dist.loc["90%"]]
len(posts)

# calculate the distribution of user activity (number of posts)
user_activity_dist: pandas.Series = (
    posts["username"]
    .value_counts()
    .describe(percentiles=[0.25, 0.4, 0.5, 0.6, 0.75, 0.8, 0.85, 0.9, 0.95])
)
user_activity_dist

selected_percentile: str = "75%"

sampled_posts: pandas.DataFrame = (
    posts[
        # select only the 80% most active users
        posts["username"].isin(
            list(
                (
                    posts["username"].value_counts()
                    > user_activity_dist.loc[selected_percentile]
                )
                .pipe(lambda _s: _s[_s])
                .index
            )
        )
    ]
    .groupby("username")
    # sample by number of least active 80%' user posts
    .sample(n=int(user_activity_dist.loc[selected_percentile]), random_state=SEED)
)

len(sampled_posts), len(sampled_posts["username"].unique())

sampled_posts

(
    sampled_posts.reset_index()[
        [
            "id",
            "author_id",
            "conversation_id",
            "first_name",
            "last_name",
            "username",
            "party",
            "created_at",
            "text",
        ]
    ].to_csv(f"{DATA_PATH}/interim/twitter.german.posts.sampled.csv")
)
