==> pipeline/preparation/de/03--User-Replies-Processing.ipynb <==
import pandas
import seaborn

import cltrier_lib

import twon_agents

SEED: int = 1
DATA_PATH: str = "../../../data"

post_ids: pandas.DataFrame = pandas.read_csv(
    f"{DATA_PATH}/interim/twitter.german.posts.csv", index_col=0
)[["id"]]
post_ids

replies: pandas.DataFrame = (
    pandas.read_csv(
        f"{DATA_PATH}/raw/twitter.german/GermanyReplies2023.csv", index_col=0
    )
    .pipe(twon_agents.data.filter_tweets, remove_w_mentions=False)
    .merge(post_ids, left_on="conversation_id", right_on="id")
    .rename(columns={"id_x": "id"})
)
replies

# calculate the distribution of user activity (number of posts)
user_activity_dist: pandas.Series = (
    replies["author_id"]
    .value_counts()
    .describe(
        percentiles=[
            0.25,
            0.5,
            0.75,
            0.8,
            0.85,
            0.9,
            0.95,
            0.96,
            0.97,
            0.98,
            0.99,
            0.995,
        ]
    )
)
user_activity_dist

selected_percentile: str = "99%"

sampled_replies: pandas.DataFrame = (
    replies[
        # select only the 40% most active users
        replies["author_id"].isin(
            list(
                (
                    replies["author_id"].value_counts()
                    > user_activity_dist.loc[selected_percentile]
                )
                .pipe(lambda _s: _s[_s])
                .index
            )
        )
    ]
    .groupby("author_id")
    # sample by number of least active 40%' user posts
    .sample(n=int(user_activity_dist.loc[selected_percentile]), random_state=SEED)
)

len(sampled_replies), len(sampled_replies["author_id"].unique())

sampled_replies

(
    sampled_replies
    # remove linebreaks and mentions at beginning
    .assign(
        text=(
            sampled_replies["text"]
            .replace(r"\n", " ", regex=True)
            .replace(r"^@(\w){1,15}\s?(@(\w){1,15})*", "", regex=True)
        )
    )[["id", "author_id", "conversation_id", "created_at", "text"]].to_csv(
        f"{DATA_PATH}/interim/twitter.german.replies.csv"
    )
)

classified_data = pandas.json_normalize(
    [
        sample.model_dump()
        for sample in cltrier_lib.classify.Pipeline()(
            list(sampled_replies["text"]), threshold=0.0
        )
    ]
)
classified_data
