==> pipeline/preparation/de/05--Dataset-Reply-Decision.ipynb <==
import random
import itertools

import pandas

DATA_PATH: str = "../../../data"
DATASET_PATH: str = f"{DATA_PATH}/processed/twitter.german.dataset.enriched.csv"

data = pandas.read_csv(DATASET_PATH, index_col=[0])
data.head()

positive_samples = {
    author: [
        sample for sample in itertools.combinations(group["text_post"].tolist(), 3)
    ]
    for author, group in data.groupby("author_id_reply")
}
sum(len(group) for group in positive_samples.values())

negative_samples = {
    author: [
        (
            *random.choice(positive_samples[author])[:2],
            random.choice(
                positive_samples[
                    random.choice(list(set(positive_samples.keys() - set([author]))))
                ]
            )[2],
        )
        for _ in range(450)
    ]
    for author in set(positive_samples.keys())
}
sum(len(group) for group in negative_samples.values())

dataset = pandas.concat(
    [
        pandas.DataFrame(
            [
                dict(
                    history_1=sample[0],
                    history_2=sample[1],
                    stimulus=sample[2],
                    action=label,
                )
                for group in sets
                for sample in group
            ]
        ).sample(150_000)
        for label, sets in [
            (True, positive_samples.values()),
            (False, negative_samples.values()),
        ]
    ]
)

dataset.to_csv(DATASET_PATH.replace("enriched", "decision"))
len(dataset)
